+++ lib/livingstyleguide/filters/html_head.rb
@@ -1,4 +1,0 @@
-LivingStyleGuide.add_filter :title do |arguments, options, block|
-  document.title = arguments.first
-  nil
-end